requirements.txt
google-api-python-client>=2.137.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
slack_sdk>=3.33.2
pydantic>=2.8.2
python-dateutil>=2.9.0.post0
pytz>=2024.1
pandas>=2.2.2
httpx>=0.27.0
requests>=2.32.3

